refactor: split ops into a dir that consists of files of three levels
