+++ python/stick/__init__.py
@@ -1,4 +1,3 @@
-from . import ops
 from . import init
 from . import data
 from . import nn
