+++ python/stick/ops/__init__.py
@@ -1,0 +1,3 @@
+from .op import *
+from .tensor_op import *
+from .tuple_op import *
